feat(replica): make NVMe-oF replica timeouts tunable via env vars
Convert hard-coded replicaCtrlrLossTimeoutSec / replicaReconnectDelaySec /
replicaFastIOFailTimeoutSec / replicaTransportAckTimeout / replicaKeepAliveTimeoutMs
from consts to vars, overridable at IM startup via LONGHORN_V2_REPLICA_* env vars.
longhorn-manager's instance_manager_controller populates those env vars from the
new data-engine-replica-* Setting CRs so operators can tune them without a code
change.

Defaults are unchanged. Only types.go's const-to-var switch required explicit
int32() casts at the few existing callers that pass these values to typed
BdevNvme* RPC wrappers.

+++ pkg/spdk/util.go
@@ -164,7 +164,7 @@ func connectNVMfBdevWithTransport(spdkClient *spdkclient.Client, controllerName,
 				spdkTransport,
 				spdktypes.NvmeAddressFamilyIPv4,
 				int32(ctrlrLossTimeout),
-				replicaReconnectDelaySec,
+				int32(replicaReconnectDelaySec),
 				int32(fastIOFailTimeoutSec),
 				replicaMultipath,
 			)
@@ -220,7 +220,7 @@ func attemptTCPFallback(spdkClient *spdkclient.Client, controllerName, ip, port 
 		spdktypes.NvmeTransportTypeTCP,
 		spdktypes.NvmeAddressFamilyIPv4,
 		int32(ctrlrLossTimeout),
-		replicaReconnectDelaySec,
+		int32(replicaReconnectDelaySec),
 		int32(fastIOFailTimeoutSec),
 		replicaMultipath,
 	)
